\name{amit}
\docType{data}
\alias{amit}
\title{Amitriptyline Dataset}
\usage{
data(amit)
}
\description{
  This dataset is from a study on the side effects of amitriptyline, which is a
drug some physicians prescribe as an antidepressant.
}
\format{This data frame consists of 7 variables on 17 subjects:
\itemize{
  \item{\code{TOT}}{ The subject's total TCAD plasma level.}
  \item{\code{AMI}}{ The amount of amitriptyline present in the
TCAD plasma level.}
  \item{\code{GEN}}{ The subject's gender, where 0 is for a male subject and 1 is for a female subject.}
  \item{\code{AMT}}{ Amount of the drug taken at
the time of overdose.}
  \item{\code{PR}}{ The subject's PR wave measurement.}
  \item{\code{DIAP}}{ The subject's diastolic blood pressure.}
  \item{\code{QRS}}{ The subject's QRS wave measurement.}
}
}
\source{Rudorfer, M. V. (1982), Cardiovascular Changes and Plasma Drug Levels After
Amitriptyline Overdose, \emph{Journal of Toxicology - Clinical Toxicology},
  \bold{19}, 67--78.
}
\references{
  Young, D. S. (2017), \emph{Handbook of Regression Methods}, CRC Press.
}
\keyword{datasets}
