\name{auditory}
\docType{data}
\alias{auditory}
\title{Auditory Discrimination Dataset}
\usage{
data(auditory)
}
\description{
  This dataset is from a study to assess auditory differences between environmental sounds given several other factors.
}
\format{This data frame consists of 3 variables on 20 subjects:
\itemize{
  \item{\code{pre.test}}{ The subject's pre-test score.}
  \item{\code{gain}}{ The gain in auditory score between the pre- and post-test administration of the treatment.}
  \item{\code{Culture}}{ A cultural status indicator of individuals, where 1 means the subject is from a "culturally-nondeprived" group and 0 means the subject is from a "culturally-deprived" group.}
}
}
\source{Hendrix, L. J., Carter, M. W., and Scott, D. T. (1982),  Covariance Analysis with
Heterogeneity of Slopes in Fixed Models, \emph{Biometrics},
  \bold{38}, 226--252.
}
\references{
  Young, D. S. (2017), \emph{Handbook of Regression Methods}, CRC Press.
}
\keyword{datasets}
