\name{Auto}
\docType{data}
\alias{Auto}
\title{Canadian Auto Insurance Dataset}
\usage{
data(Auto)
}
\description{
  This dataset is from The Statistical Unit of the Canadian Underwriters' Association collated automobile insurance policies (policy years 1956 and 1957) for private passenger automobile liability for non-farmers in Canada excluding those in the province of Saskatchewan.
}
\format{This data frame consists of 20 categories (rows) and 6 variables (columns):
\itemize{
  \item{\code{Merit}}{ Merit rating of policyholder.}
  \item{\code{Class}}{ Class rating of policyholder.}
  \item{\code{Insured}}{ Earned car years under the policy.}
  \item{\code{Premiums}}{ Earned premium at present rates (in 1000's of Canadian dollars).}
  \item{\code{Claims}}{ Number of claims incurred.}
  \item{\code{Cost}}{ Incurred losses (in 1000's of Canadian dollars).}
}
}
\source{Bailey, R. A. and Simon, L. J. (1960),   Two Studies in Automobile Insurance Ratemaking, \emph{ASTIN Bulletin},
  \bold{1}, 192--217.
}
\references{
  Young, D. S. (2017), \emph{Handbook of Regression Methods}, CRC Press.
}
\keyword{datasets}
