\name{BAC}
\docType{data}
\alias{BAC}
\title{Blood Alcohol Concentration Dataset}
\usage{
data(BAC)
}
\description{
  This dataset is from a study to compare the blood alcohol concentration (BAC) of subjects using two different methods.
}
\format{This data frame consists of 2 variables measured on 15 subjects:
\itemize{
  \item{\code{breath}}{ BAC obtained using the Breathalyzer Model 5000.}
  \item{\code{labtest}}{ BAC based on a breath estimate in a laboratory.}
}
}
\source{Krishnamoorthy, K., Kulkarni, P. M., and Mathew, T. (2001), Multiple Use One-Sided Hypotheses Testing in Univariate Linear Calibration, \emph{Journal of Statistical Planning and Inference},
  \bold{93}, 211--223.
}
\references{
  Young, D. S. (2017), \emph{Handbook of Regression Methods}, CRC Press.
}
\keyword{datasets}
