\name{cheese}
\docType{data}
\alias{cheese}
\title{Cheese-Tasting Experiment Dataset}
\usage{
data(cheese)
}
\description{
  This dataset is from an experiment concerning the effect on taste of various cheese additives.
}
\format{This data frame (36 rows by 3 columns) is a tabulation of the responses by 208 subjects to 4 different cheeses:
\itemize{
  \item{\code{Cheese}}{ The cheese additive used (four levels labeled A, B, C, and D).}
  \item{\code{Response}}{ The response based on the 9-point hedonic scale.}
  \item{\code{N}}{ The number of subjects who responded according to the value in \code{Response} for the cheese additive in \code{Cheese}.}  
}
}
\source{McCullagh, P. and Nelder, J. A. (1989), \emph{Generalized Linear Models}, CRC Press.
}
\references{
  Young, D. S. (2017), \emph{Handbook of Regression Methods}, CRC Press.
}
\keyword{datasets}
