\name{chem}
\docType{data}
\alias{chem}
\title{Odor Dataset}
\usage{
data(chem)
}
\description{
  This dataset is from an experiment that was designed to determine the effects of three factors in reducing the unpleasant odor in a chemical product being sold for household use.
}
\format{This data frame consists of 4 variables (stored in coded form) at 15 design points:
\itemize{
  \item{\code{odor}}{ A measure of the chemical's odor.}
  \item{\code{temp}}{ Temperature at time of measurement - uncoded units are 40, 80, and 120.}
  \item{\code{ratio}}{ Gas-liquid ratio - uncoded units are 0.3, 0.5, and 0.7.}
  \item{\code{height}}{ Packing height - uncoded units are 2, 4, and 6.}
}
}
\source{John, P. W. (1971), \emph{Statistical Design and Analysis of Experiments}, MacMillan Company.
}
\references{
  Young, D. S. (2017), \emph{Handbook of Regression Methods}, CRC Press.
}
\keyword{datasets}
