\name{compasst}
\docType{data}
\alias{compasst}
\title{Computer-Assisted Learning Dataset}
\usage{
data(compasst)
}
\description{
  This dataset is from a study of computer-assisted learning by students in an effort to assess the cost of computer time.

}
\format{This data frame consists of 2 variables measured on 12 students:
\itemize{
  \item{\code{num.responses}}{ Total number of responses in completing a lesson.}
  \item{\code{cost}}{ Cost of the computer time (in cents).}
	}
}
\source{Kutner, M. H., Nachtsheim, C. J., Neter, J., and Li, W. (2005) \emph{Applied Linear Statistical Models, Fifth Edition}, McGraw-Hill/Irwin.
}
\references{
  Young, D. S. (2017), \emph{Handbook of Regression Methods}, CRC Press.
}
\keyword{datasets}
