\name{cracker}
\docType{data}
\alias{cracker}
\title{Cracker Dataset}
\usage{
data(cracker)
}
\description{
  This dataset is from marketing research on the sales of crackers for a particular company.
}
\format{This data frame consists of 15 stores, each receiving a particular promotional strategy, with the following 4 variables (columns):
\itemize{
  \item{\code{treat}}{ An indicator for which of the three marketing strategies (treatment) was employed.}
  \item{\code{store}}{ The store number within the particular treatment.}
  \item{\code{y}}{ The number of cases of the crackers sold during the promotional period.}
  \item{\code{x}}{ The store's cracker sales during the preceding sales period.}
	}
}
\source{Kutner, M. H., Nachtsheim, C. J., Neter, J., and Li, W. (2005) \emph{Applied Linear Statistical Models, Fifth Edition}, McGraw-Hill/Irwin.
}
\references{
  Young, D. S. (2017), \emph{Handbook of Regression Methods}, CRC Press.
}
\keyword{datasets}
