\name{fiber}
\docType{data}
\alias{fiber}
\title{Fiber Strength Dataset}
\usage{
data(fiber)
}
\description{
  This dataset is from a study about the strength of a particular type of fiber based on the amount of pressure applied.
}
\format{This data frame consists of 30 samples with the following 2 variables measured:
\itemize{
  \item{\code{pressure}}{ The amount of water pressure applied (measured in bars); the unique levels are 60, 80, 100, 120, 150, and 200.}
  \item{\code{tensile}}{ The tensile strength of fiber (measured in N/5 cm).}
}
}
\source{Ndaro, M. S., Jin, X.-Y., Chen, T., and Yu, C.-W. (2007),  Splitting of Islands-in-the-Sea Fibers (PA6/COPET) During Hydroentangling of Nonwovens, \emph{Journal of Engineered Fibers and Fabrics},
  \bold{2}, 1--9.
}
\references{
  Young, D. S. (2017), \emph{Handbook of Regression Methods}, CRC Press.
}
\keyword{datasets}
