\name{fly}
\docType{data}
\alias{fly}
\title{Fruit Fly Dataset}
\usage{
data(fly)
}
\description{
  This dataset is from a study on the effects of temperature on development of the common fruit fly.
}
\format{This data frame consists of 23 batches with the following 8 variables measured:
\itemize{
  \item{\code{temp}}{ The experimental temperature (in degrees Celsius).}
  \item{\code{exp.no}}{ The experiment number.}
  \item{\code{duration}}{ The mean duration of the embryonic period (in hours).}
  \item{\code{dur.var}}{ The standard deviation of the duration of the recorded embryonic period.}
  \item{\code{batch}}{ The number of eggs in each batch.}
  \item{\code{batch.sd}}{ The standard deviation of the number of eggs in each batch.}
  \item{\code{egg.temp}}{ The temperature in which the eggs were laid (in degrees Celsius).}
  \item{\code{egg.dur}}{ The duration the eggs remained in the given conditions (in hours).}
}
}
\source{Powsner, L. (1935), The Effects of Temperature on the Durations of the Developmental Stages of \emph{Drosophila Melanogaster}, \emph{Physiological Zoology},
  \bold{8}, 474--520.
}
\references{
  McCullagh, P. and Nelder, J. A. (1989), \emph{Generalized Linear Models}, CRC Press.

  Young, D. S. (2017), \emph{Handbook of Regression Methods}, CRC Press.
}
\keyword{datasets}
