\name{gas}
\docType{data}
\alias{gas}
\title{Natural Gas Prices Dataset}
\usage{
data(gas)
}
\description{
  This dataset is of the monthly observations of spot prices for
natural gas from January 1988 to October 1991 for the states of Louisiana
and Oklahoma.
}
\format{This data frame consists of a total of 46 (monthly) observations of spot prices for the 2 states stated above:
\itemize{
  \item{\code{OK}}{ Oklahoma spot prices for natural gas (dollars per million British thermal
units).}
  \item{\code{LA}}{ Louisiana spot prices for natural gas (dollars per million British thermal
units).}
}
}
\source{Wei, W. W. S. (2005), \emph{Time Series Analysis: Univariate and Multivariate Methods}, Pearson.

}
\references{
  Young, D. S. (2017), \emph{Handbook of Regression Methods}, CRC Press.
}
\keyword{datasets}
