\name{GRB}
\docType{data}
\alias{GRB}
\title{Gamma-Ray Burst Dataset}
\usage{
data(GRB)
}
\description{
  This dataset consists of measurements on gamma-ray bursts, which are short, intense flashes of gamma-ray radiation that occur at (seemingly) random times and
locations in space. 
}
\format{This data frame consists of 63 measurements of the following 2 variables:
\itemize{
  \item{\code{TIME}}{ Time of measurement (in seconds).}
  \item{\code{FLUX}}{ X-ray flux measurement (in units of 10^(-11) erg/cm2/s, 2-10 keV).}
}
}
\source{Blustin, A. J., Band, D., Barthelmy,  S., Boyd,  P., Capalbi, M., Holland, S. T., Marshall, F. E., Mason, K. O., Perri, M., Poole, T., Roming, P., Rosen, S., Schady, P., Still, M., Zhang, B., Angelini, L., Barbier, L., Beardmore, A., Breeveld, A., Burrows, D. N., Cummings, J. R., Canizzo, J., Campana, S., Chester M. M., Chincarini, G., Cominsky, L. R., Cucchiara, A., de Pasquale, M., Fenimore, E. E., Gehrels, N., Giommi, P., Goad, M., Gronwall, C., Grupe, D., Hill, J. E., Hinshaw, D., Hunsberger, S., Hurley K. C., Ivanushkina, M., Kennea, J. A., Krimm, H. A., Kumar, P., Landsman, W., La Parola, V., Markwardt, C. B., McGowan, K., Meszaros, P., Mineo, T.,  Moretti, A., Morgan, A., Nousek, J., O'Brien, P. T., Osborne, J. P., Page, K., Page, M. J., Palmer, D. M., Parsons, A. M., Rhoads, J., Romano, P., Sakamoto, T., Sato, G., Tagliaferri, G., Tueller, J., Wells, A. A. and White, N. E. (2006),  Swift Panchromatic Observations of the Bright Gamma-Ray Burst GRB 050525a, \emph{The Astrophysical Journal},
  \bold{637}, 901--913.
}
\references{
  Young, D. S. (2017), \emph{Handbook of Regression Methods}, CRC Press.
}
\keyword{datasets}
