\name{HoRM-package}
\alias{HoRM-package}
\alias{HoRM}
\docType{package}
\title{
Supplemental Functions and Datasets for "Handbook of Regression Methods"
}
\description{
Various wrapper functions and datasets to supplement examples for the book "Handbook of Regression Methods" by D. S. Young.
}
\details{
\tabular{ll}{
Package: \tab HoRM\cr
Type: \tab Package\cr
Version: \tab 0.1.2\cr
Date: \tab 2019-04-24\cr
Imports: \tab ggplot2, MASS, orthopolynom, quantmod, rsm, stats4\cr
License: \tab GPL (>= 2)\cr
}
}
\author{
Derek S. Young, Ph.D.

Maintainer: Derek S. Young <derek.young@uky.edu>
}
\references{
  Young, D. S. (2017), \emph{Handbook of Regression Methods}, CRC Press.
}

