\name{JamesBond}
\docType{data}
\alias{JamesBond}
\title{James Bond Dataset}
\usage{
data(JamesBond)
}
\description{
  This dataset consists of various metrics pertaining to the officially-produced \emph{James Bond} films.
}
\format{This data frame consists of 18 variables measured on the 24 films:
\itemize{
  \item{\code{Year}}{ Year of the film's theatrical release.}
  \item{\code{Movie}}{ Title of the movie.}
  \item{\code{Bond}}{ Actor who played James Bond.}
  \item{\code{US_Gross}}{ The film's U.S. gross (in U.S. dollars).}
  \item{\code{US_Adj}}{ The film's 2013-adjusted U.S. gross (in 1000's of U.S. dollars).}
  \item{\code{World_Gross}}{ The film's worldwide gross (in U.S. dollars).}
  \item{\code{World_Adj}}{ The film's 2013-adjusted worldwide gross (in 1000's of U.S. dollars).}
  \item{\code{Budget}}{ The film's budget (in U.S. dollars).}
  \item{\code{Budget_Adj}}{ The film's 2013-adjusted budget (in 1000's of U.S. dollars).}
  \item{\code{Film_Length}}{ Length of the theatrical release.}
  \item{\code{Avg_User_IMDB}}{ The average user rating on IMDB (www.imdb.com).}
  \item{\code{Avg_User_Rtn_Tom}}{ The average user rating on Rotten Tomatoes (www.rottentomatoes.com).}
  \item{\code{Conquests}}{ The number of "conquests" by Bond in the film.}
  \item{\code{Martinis}}{ The number of martinis Bond drank in the film.}
  \item{\code{BJB}}{ The number of times Bond stated "Bond. James Bond." in the movie.}
  \item{\code{Kills_Bond}}{ The number of people killed by Bond.}
  \item{\code{Kills_Others}}{ The number of people killed in the film by people other than Bond.}
  \item{\code{Top_100}}{ An indicator where a value of 1 means the title song  within the top 100 on the UK Singles Chart and the U.S. Billboard Hot 100 and a value of 0 means it did not.}
}
}
\source{Young, D. S. (2014), Bond. James Bond. A Statistical Look at Cinema's Most Famous Spy, \emph{CHANCE},
  \bold{27}(2), 21--27.

Young, D. S. (2019), Bond. James Bond. A Statistical Look at Cinema's Most Famous Spy (\emph{The Best of CHANCE Issue}), \emph{Chance},
  \bold{32}(1), 27--35.
}
\references{
  Young, D. S. (2017), \emph{Handbook of Regression Methods}, CRC Press.
}
\keyword{datasets}
