\name{light}
\docType{data}
\alias{light}
\title{Light Dataset}
\usage{
data(light)
}
\description{
  This dataset is from an experiment where light was transmitted through a chemical solution and an optical reading was recorded.
}
\format{This data frame consists of 2 variables measured on 12 different instances:
\itemize{
  \item{\code{reading}}{ The optical reading.}
  \item{\code{concentration}}{ The concentration of the chemical.}
}
}
\source{Graybill, F. A. and Iyer, H. K. (1994), \emph{Regression Analysis: Concepts and Applications}, Duxbury Press.
}
\references{
  Young, D. S. (2017), \emph{Handbook of Regression Methods}, CRC Press.
}
\keyword{datasets}
