\name{repair}
\docType{data}
\alias{repair}
\title{Computer Repair Dataset}
\usage{
data(repair)
}
\description{
  This dataset is from a random sample of service
call records for a computer repair company.
}
\format{This data frame consists of a sample of 14 companies with the following 2 variables measured:
\itemize{
  \item{\code{minutes}}{ The  length of service call (in minutes).}
  \item{\code{units}}{ The number of components repaired or replaced during the service call.}
}
}
\source{Chatterjee, S. and Hadi, A. S. (2012), \emph{Regression Analysis by Example}, John Wiley and Sons, Inc.
}
\references{
  Young, D. S. (2017), \emph{Handbook of Regression Methods}, CRC Press.
}
\keyword{datasets}
