\name{tortoise}
\docType{data}
\alias{tortoise}
\title{Tortoise Eggs Dataset}
\usage{
data(tortoise)
}
\description{
  This dataset is from a study on the number
of eggs in female gopher tortoises in southern Florida.
}
\format{This data frame consists of 2 variables measured on 18 tortoises:
\itemize{
  \item{\code{length}}{ The carapace length (in millimeters).}
  \item{\code{clutch}}{ The number of eggs (clutch size).}
}
}
\source{Ashton, K. G., Burke, R. L., and Layne, J. N. (2007),  Geographic Variation in Body
and Clutch Size of Gopher Tortoises, \emph{Copeia},
  \bold{2007}, 355--363.
}
\references{
  Young, D. S. (2017), \emph{Handbook of Regression Methods}, CRC Press.
}
\keyword{datasets}
