\name{toy}
\docType{data}
\alias{toy}
\title{Toy Dataset}
\usage{
data(toy)
}
\description{
  A made-up (toy) dataset.
}
\format{This data frame consists of 2 made-up variables for a sample of size 5:
\itemize{
  \item{\code{x}}{ The made-up x values.}
  \item{\code{y}}{ The made-up y values.}
}
}
\source{  Young, D. S. (2017), \emph{Handbook of Regression Methods}, CRC Press.
}
\references{
  Young, D. S. (2017), \emph{Handbook of Regression Methods}, CRC Press.
}
\keyword{datasets}
