\name{wood}
\docType{data}
\alias{wood}
\title{Pulp Property Dataset}
\usage{
data(wood)
}
\description{
  This dataset is from a study about the pulp properties of wood density of the Australian
blackwood tree.
}
\format{This data frame consists of 2 variables measured on 7 samples:
\itemize{
  \item{\code{pulp}}{ The percentage of pulp yield.}
  \item{\code{Kappa}}{ The Kappa number, which is a measurement of standard potassium permanganate
solution that the pulp will consume.}
}
}
\source{Santos, A., Anjos, O., Amaral, M. E., Gil, N., Pereira, H., and Simoes, R. (2012),  Influence on Pulping Yield and Pulp Properties of Wood Density of \emph{Acacia
melanoxylon}, \emph{Journal of Wood Science},
  \bold{58}, 479--486.
}
\references{
  Young, D. S. (2017), \emph{Handbook of Regression Methods}, CRC Press.
}
\keyword{datasets}
