\name{yarn}
\docType{data}
\alias{yarn}
\title{Yarn Fiber Dataset}
\usage{
data(yarn)
}
\description{
  This dataset is from a mixture experiment regarding a fiber blend that
is spun into yarn to make draperies.}
\format{This data frame consists of 4 variables measured at 15 design points for a \{3,2\} simplex lattice design:
\itemize{	
  \item{\code{x1}}{ The proportion of the polyethylene component.}
  \item{\code{x2}}{ The proportion of the polystyrene component.}
  \item{\code{x3}}{ The proportion of the polypropylene component.}
  \item{\code{y}}{ The measurement of yarn elongation.}
}
}
\source{Cornell, J. A. (2002), \emph{Experiments with Mixtures: Designs, Models, and the Analysis of Mixture Data, Third Edition}, John Wiley and Sons, Inc.

}
\references{
  Young, D. S. (2017), \emph{Handbook of Regression Methods}, CRC Press.
}
\keyword{datasets}
